backend/requirements.txt
Django==4.0
djangorestframework==3.12.4
torch==1.10.0
transformers==4.16.2
pandas==1.3.5

Explanation: This file lists the Python packages required to run the backend application. You can install these packages using pip install -r requirements.txt.